Add Rake task to import activities
This adds a Rake task to run an Activity import with the command:

```
rake activites:import CSV=path/to/file.csv ORGANISATION_ID=8c3b69ec-1e9c-49ae-8e04-7c5d3826b253
```

This parses the CSV from the path and runs the importer. If the import is
successful, the number of added / updated Activities is presented to the user. If
there are any errors, these are played back to the user.

I've kept the tests pretty light touch, and only testing the code that gets introduced
in the Rake task, as we don't want a whole bunch of CSV files clogging up the
test suite, and we'd only be repeating ourselves anyway, as the everything is
tested in the `ImportFromCsv` class.