`intended_beneficiaries` is now always optional
In order to facilitate Q3 reporting, we are trying to simulate the behaviour
of the trackers on this geography/intended beneficiaries step, so users will be
able to report their benefitting countries as they would do on the trackers,
without encountering validation problems. For this, now a user can decide
whether or not they add intended beneficiaries even if they have selected a
region as recipient. This is to reflect some of the trackers we've seen in which
regions are selected but no intended beneficiaries are added.