Infer Geography field
When filling out the form, a Geography field is required. This is not a column that's
provided in the CSV, but we can infer what it would be based on the presence or
otherwise of the recipient_region field. If it's present, the Geography is the region,
otherwise it's the country.