Sort same-quarter reports by their creation time
In storing the history for planned disbursements, we have linked all
disbursement records to a report, and assumed reports can be totally
ordered by their financial quarter.

However, sometimes we will have two reports for the same fund and
organiation that fall in the same quarter, for procedural reasons. We
break the ordering tie on such records using their `created_at` field.

Sorting data causally based on the time it happened to be written is not
especially robust -- it still allows two reports to be created at the
same second, or in reverse order compared to the financial quarters they
relate to. In the absence of a more robust causal model I am therefore
only using created_at to break ties in cases where two reports fall in
the same quarter -- not using created_at to sort the whole set.

This feels relatively low-risk to me, but it remains _critical_ the the
reports for a fund and organisation are totally ordered -- there is only
ever one active report, and once it's approved we open the next one. We
don't have active reports that overlap in time and we don't go back and
modify approved reports -- doing that would break the causal assumptions
here.