Prevent forecasts about the past being created
When a forecast for an activity and quarter is given, it must related to
a quarter in the future. Giving forecasts about the present or the past
makes no sense and may violate causality. Therefore we reject forecasts
if their quarter is not in future relative to the current reporting
quarter.