Remove error layout and navigation
The change to check the presence of the `current_user` instance variable
allows the shared navigation partial to work everywhere which also means
we no longer need the error layout or navigation.

As the navigation is not shown at all to users who are not signed in, we
never show the sign in link to them so this is redundant.

I've added a link to the error pages to the root of the application in
case a signed out user does see an error and does not know to click the
site title to return, this should help stop errors being dead ends.