Make a planned disbursement's activity and quarter read-only
The activity and quarter a PlannedDisbursement relates to must not be
modified -- under our model of history, these are now part of the
object's identity and place it in the historical log for that activiy
and quarter. Changing these fields would need to nonsensical history
data being generated, so we make them read-only.

We must also not change the report a PlannedDisbursement is linked to as
that would change its position in the history, and again lead to the log
making no sense.