Relax the policy for editing planned disbursements
When a user edits a planned disbursement in RODA, they may not be
creating or updating a specific database record. The
PlannedDisbursementHistory service _may_ create or update records
depending on the current state and how it relates to the current report.

Therefore, being allowed to edit a planned disbursement only requires
that there's a current report. It does _not_ require that disbursement
to belong to the current report -- if it does not, a new disbursement
record will be created by PlannedDisbursementHistory.