Put unique indexes on the identities of planned disbursements
Storing revision history for planned disbursements requires a few
assumptions to be upheld about our data and I would like to enforce
those in the database, to the extent possible.

One of these assumptions is that every `planned_disbursements` record
for a given activity and quarter, belongs to a different report. Having
two records for the same report in the same history series is ambiguous,
so we'd like to prevent that state occurring.

There are two cases to consider:

- At level A/B, planned disbursements are not linked to reports -- for a
  given activity and quarter there must be at most one original record
  and at most one revised record. i.e. if `report_id` is null then
  `planned_disbursement_type` must be unique.

- At level C/D, planned disbursements are linked to reports, and each
  record for a given activity and quarter should have a different
  `report_id`.

These indexes enforce these two constrains by putting UNIQUE constraints
on these partial indexes:

    (
      parent_activity_id,
      financial_year,
      financial_quarter,
      planned_disbursement_type
    ) WHERE report_id IS NULL

    (
      parent_activity_id,
      financial_year,
      financial_quarter,
      report_id
    ) WHERE report_id IS NOT NULL

This required updating one test that was creating two forecasts for the
same report.

All the data in production currently conforms to these constrainsts so
this should run fine when deployed.