Intended beneficiaries
This splits the intended beneficiaries by the `;` delimiter and checks they are in
the correct format.

I've also added the inferred region to the attributes once all the validation has
happened. This happens after the geography has been calculated, as we need to
check for the presence / absence of the region first.