Activity objectives added to the activity form
Aims/Objectives is one of the additional fields we need to implement before Q3
reporting. This field is captured after the activity description and it is
mandatory for levels B, C, and D activities. It is a text area with no character
limit.

This field is reported to IATI and it is also present on the CSV report. We will
see that logic implemented on further commits.