Activity objectives gets reported to IATI
The aims/objectives of an activity need to be reported to IATI under the field
`description, type 2`. This field is only mandatory on levels B, C, and D so
we won't include it in the XML for Funds.

These changes have been checked on the IATI validator, having the files passing
green.