Add `gcrf_challenge_area` column to the Activity model
This attribute is non-standard, as it doesn't have an obvious key-value
pair, so originally I stored this as a text field only. This risks data
integrity problems if any of the radio button text is changed in the
future, as this would've been stored directly in the model.

I therefore decided to store it as an integer, with each possible option
mapped in the order they appear, ensuring that "N/A" is assigned to `0`.

Additionally, the database column has `null: true` because we need to
store the difference between "this attribute is not applicable" and
"this attribute has not been supplied (or does not apply)".