Correct links between forecasts and reports
Reviewing the production data around planned disbursements, we found a
few anomalies where things didn't make sense. Forecasts related to level
A/B activities should not be linked to reports -- those activities
belong to BEIS and don't go through the reporting cycle. Forecasts for
level C/D activities _should_ be linked to a report. But we see a number
of records that don't reflect these rules:

    > pds = PlannedDisbursement.includes(parent_activity: [:organisation], report: [:fund, :organisation])

    > pds.where(activities: { level: %w[fund programme] }).where.not(report_id: nil).count
    => 92
    > pds.where(activities: { level: %w[project third_party_project] }).where(report_id: nil).count
    => 3

There are 92 level A/B forecasts linked to reports. The reports these
records are linked to, are:

    > pds.where(activities: { level: %w[fund programme] }).where.not(report_id: nil)
        .map(&:report).uniq.map { |r| [r.description, r.financial_quarter, r.financial_year, r.fund.roda_identifier, r.organisation.name] }

    => [["Historic tracker data (GCRF)", nil, nil, "GCRF", "Royal Society"],
        ["Historic tracker data (GCRF)", nil, nil, "GCRF", "Royal Academy of Engineering"],
        ["Historic tracker data (GCRF)", nil, nil, "GCRF", "British Academy"]]

So all these reports are historic imports, not open reports that
delivery partners are filling out. It seems fine to unlink the level A/B
forecasts from these by setting `report_id = NULL` on those records.

There are 3 level C/D forecasts not linked to reports, so we need to
find a report to link them to. For each one, there is currently only a
single report matching the forecast's parent activity's fund and
organisation. In fact, it's the same report for all three:

    > no_report = pds.where(activities: { level: %w[project third_party_project] }).where(report_id: nil)

    > no_report.map { |r| Report.where(fund: r.parent_activity.associated_fund, organisation: r.parent_activity.organisation)
        .map { |r| [r.id, r.description, r.financial_quarter, r.financial_year] } }

    => [[["92d8deff-8f2b-4573-b02c-0655cea59c48", "Historic tracker data (Newton fund)", nil, nil]],
        [["92d8deff-8f2b-4573-b02c-0655cea59c48", "Historic tracker data (Newton fund)", nil, nil]],
        [["92d8deff-8f2b-4573-b02c-0655cea59c48", "Historic tracker data (Newton fund)", nil, nil]]]

Again, these are historic reports, not ones being actively filled out by
delivery partners. This looks like a link that was missed during ingest
and it would be fine to link these 3 forecasts to the appropriate
report.

In the migration, we check that only one report matches the given search
criteria in case these assumptions change by the time this migration is
run.

As a side effect, this change solves another inconsistency in the data.
There are 92 forecasts where the parent activity and report belong to
different organisations. All these forecasts belong to level B
activities:

    > pds.where.not(report_id: nil).select { |r| r.parent_activity.organisation != r.report.organisation }.size
    => 92
    > pds.where.not(report_id: nil).select { |r| r.parent_activity.organisation != r.report.organisation }.map { |r| r.parent_activity.level }.uniq
    => ["programme"]

Setting `report_id = NULL` on these forecasts will remove these
inconsistencies since these forecasts will no longer belong to a report
at all.