Enforce a limit of one "historic" report per series
To make the historical record of planned disbursements make sense, we
need to ensure that the series of reports belonging to a fund and
organisation is _totally ordered_. This ordering is defined over the
financial quarter each report belongs to.

However, some reports have null for their `financial_quarter` and
`financial_year` fields. These reports represent the historical data we
imported for an organisation, before they began using RODA. There should
be only one such report for each fund and organisation, and it should
sort before all other quarterly reports in the history.

To ensure this remains the case, we add a partial index to put a unique
constraint on `(fund_id, organisation_id)` where either
`financial_quarter` or `financial_year` is null.

The reports that exist in production already obey this constraint and so
the migration should run successfully:

    > reports = Report.where(financial_quarter: nil).or(Report.where(financial_year: nil))

    > reports.count
    => 11
    > reports.map { |r| [r.fund_id, r.organisation_id] }.uniq.size
    => 11