Enforce a limit of one "editable" report per series
To make the historical record of planned disbursements make sense, we
need to ensure that the series of reports belonging to a fund and
organisation is _totally ordered_. This ordering is defined over the
financial quarter each report belongs to.

Part of how this sequence is maintained, is that at any given time there
is only one editable report -- one whose state is `active` or
`awaiting_changes` -- in a series. Any data entered by the delivery
partner is linked to this report. Eventually the report is approved and
no further data can be added to it. At that point, a new editable report
is created and all new data goes into that one.

It's important that old reports are not re-opened, otherwise it becomes
ambiguous which report new data belongs to, and the sequence becomes
broken. This index prevents us making a report editable if any other
editable report exists within the same (fund, organisation) series.

The data in production currently obeys this constraint and so the
migration should run cleanly:

    > Report.editable.count
    => 8
    > Report.editable.map { |r| [r.fund_id, r.organisation_id] }.uniq.size
    => 8