Use controller spec for testing step skipping
We're currently using integration tests to check that the correct steps
are shown to the user, depending on the level of the activity being
edited.

This is quite slow and gets complicated when you started dealing with
combinations of both level and fund affecting the steps being skipped.

The 'wicked' gem's `render_wizard` method redirects the user to the next
form step when `skip_step` is called. Therefore a controller test can
be used to test this logic.