Mark all "historic" reports as approved
We have a set of reports with no financial quarter; these represent the
sets of "historic" data we imported for an organisation during
onboarding, before they began using RODA. They are considered to predate
all quarterly reports, for the purposes of recording the history of our
data.

New data entered via RODA should _never_ be assigned to these reports.
Many objects like transactions and forecasts are automatically linked to
the current "editable" (active or awaiting_changes) report for that
organisation, and having these "historic" reports in an editable state
allows data entered via RODA to be linked to them.

By marking them as approved, we prevent the RODA service linking new
data to them, so we can only add data to them by explicitly choosing to
do so via an ingest process.