Prevent forecast scopes from being NULL
The "scope" of a forecast (planned disbursement) is the thing it relates
to: the object you plan to spend money on. A forecast is a plan to spent
money on a certain activity in a certain quarter, and so the fields
describing that object should not be nullable.

This migration prevents the parent_activity_id, financial_quarter, and
financial_year from being NULL. No records have NULL in these columns in
production, so this migration should deploy successfully.