Set `created_at` automatically in ReportingCycle
Some of our code and tests assume that the `created_at` field on Report
always falls within the time period that the report relates to. That
means reports created during tests need to set this field explicitly in
order for the tests to represent what the code does in production.

Previously this class incremented `created_at` by 1 second for each
report it created, just to make sure they're totally ordered. We now set
`created_at` based on the report's financial quarter so that the above
assumption holds.