Return the forecast value for a report's quarter
For calculating variances in a report, we want to know the forecast for
an activity in that report's financial quarter. In general, an activity
will have forecasts for many different quarters, but for variances we
just want to know the forecast for _that report's quarter_, in order to
compare it to the transactions in that report.

Forecast information is also versioned; the value of the forecast for an
activity and quarter may change in each report and we store that
history. In general, a report should only contain values for forecasts
relating to quarters later than the report's own quarter. In fact,
PlannedDisbursementHistory ensures this.

To get the forecast spend for a given report's quarter, we use
PlannedDisbursementOverview. Its `values_at_report` method returns the
forecasts across all quarters for an activity, with the values as they
were at the time of the given report. We just need to select the
forecast for the particular quarter we're interested in -- the quarter
the report relates to.