Adding `country_delivery_partners` to DB
`country_delivery_partners` is a new field on the Activity form. The user will
be able to add up to 10 different country delivery partners, so we will be
storing them as an array.

This field only applies for Newton funded activities, so validation has been
introduced on a later commit.