`country_delivery_partners` added to activity form
The new field Country delivery partners is now in the activity form. This field
is only applicable for Newton funded activities, and only mandatory for
programmes (level B). Optional on levels C and D, and not required (not to be
present) for Funds (model validation on next commit).

Due to the nature of the UI design for this field, we have customised the UI
elements we use to collect the data, instead of using the FormBuilder. The
elements available on FormBuilder would not give us the functionality we were
looking for.

During the development of this piece of work we also spotted an error where
activity fields, despite not complying with validation rules, were displaying
a success message. The error came from the `update_form_state` method on the
activity form controller. This is now fixed.

We also got some failing tests with a Capybara:InfiniteRedirect error. This was
due to fund activities skipping several steps on the activity form. The limit
for Capybara was set on 5. This is by default. When I added
`country_delivery_partners` to the form, since this step is also skipped on funds
we got this error, as this step was the 6th step we were skipping. We have solved
this issue by increasing the limit of redirects Capybara will accept, but it
would be worth looking at this again when we have the time and maybe use
other methods available for Wicked Wizard on the controller, like `jump_to`
instead of `skip_step`.