Get the forecast for a specific quarter from a report
In order to give the forecasts for the next four quarters following a
report, we need to extend this method for fetching the forecast for a
specific quarter, not just the quarter the report relates to. By
default, the method continues to return the forecast value for the
report's quarter, but this can be overridden via optional arguments.

We also need to call `@activity.id` rather than use `@activity` to build
a query, because we're going to be passing ActivityPresenter objects in
here to make the CSV exporter work.