Use the forecast history for the next four quarters
We're currently using Activity#forecasted_total_for_date_range to
calculate the forecasts for the next four quarters when exporting CSV.
The method uses dates rather than quarters, and isn't aware of the
version history model for planned disbursements we've implemented in
PlannedDisbursementHistory and PlannedDisbursementOverview.

Therefore we replace it with a call to
PlannedDisbursementOverview#value_for_report, passing in each quarter
and year we're interested in. Activity#forecasted_total_for_date_range
and Report#quarters_to_date_ranges methods are no longer needed and can
be deleted.