Refactor the PlannedDisbursementOverview interface
I don't like how this class is growing new methods, so that it now
supports the interface:

- latest_values
- values_at_report(report)
- value_for_report(report, financial_quarter:, financial_year:)

These don't express the use cases very well:

- `latest_values` is for getting the latest versions of forecasts for
  all quarters, for display on the "financials" page.
- `values_at_report(report)` is not used publically other than for
  testing, but it's used inside other methods to get the forecasts in
  the state they were in when a given report was approved.
- `value_for_report(report)` is for getting the forecasted spend for the
  quarter a report relates to, for the purposes of calculating
  variances.
- `value_for_report(report, financial_quarter:, financial_year:)` is for
  getting the value for a certain quarter, as recorded in a certain
  report, for the purpose of exporting some near-future forecasts in CSV
  output.

I'm not a fan of the use of optional parameters for that last case; the
caller should really be explicit about the fact it's performing a
distinct operation. So, I'm replacing the `values_at_report` and
`value_for_report` methods with this interface:

- snapshot(report).all_quarters
- snapshot(report).value_for_report_quarter
- snapshot(report).value_for(financial_quarter:, financial_year:)

This makes the use cases more distinct, and better expresses the idea
that we're fetching a "snapshot" of the forecast data at a point in
time, as captured by a particular report.