Allow a Report to be explicitly specified in PlannedDisbursementHistory
The method `PlannedDisbursementHistory#set_value` usually assigns the
PlannedDisbursement objects it creates to the current editable Report
for the fund and organisation that the relevant Activity relates to.
This is what we want in normal use, when end users are adding data via
the website.

To allow developers to back-fill some data from the previous quarter, we
need to add data to some non-editable reports from previous quarters.
These reports are the latest in their series, so this does not involve
rewriting history per se, but the reports have been closed so the
default behaviour of PlannedDisbursementHistory won't find the right
report.

Instead, we're planning on having developers upload files containing the
back-fill data, and they will specify which report to link that data to.