Import forecasts from spreadsheets
We're planning to have developers back-fill some forecast data into last
quarter's reports by providing a CSV file. That file will have a column
named "RODA identifier" that identifies the activity for each forecast,
and then columns with names like "FC 2020/21 FY Q3 (Oct, Nov, Dec)" that
contain the forecast value for the given quarter.

The job of ImportPlannedDisbursements is to interpret this format and
save the given values by extracting the financial year and quarter from
those column headings and using them to fill in the history.

The class takes a Report as input, as this data will be explicitly
assigned to a specific report rather than to the current editable one.

Zero values are ignored; any previously existing value for that activity
and quarter will remain in effect.