Check that imported forecasts go into the latest report
When developers upload files to back-fill the forecast data for last
quarter, they will specify which report to add that data to. This is
because last quarter's reports are no longer editable and so won't be
the one PlannedDisbursementHistory would normally pick.

As a safety check, we make sure the given report is still the latest one
for the fund and organisation in question. The versioing model we use
for forecasts assumes that reports and the data they contain are totally
ordered -- we do not go back and change older reports that have been
committed to the history.

It's still safe to modify the most recent report, even if it's no longer
editable by end users, as long as a new report has not been opened for
users to add data to. The data users enter in Report 2 will be informed
by the data the site shows for Report 1, so if we go back and change
Report 1 that may invalidate data in Report 2.

So to make sure what we're doing is safe, and rather than implementing a
model that supports arbitrarily editing history, we check that the given
Report is the latest one in its sequence, and is therefore safe to
modify as users have not yet begun filling a new report out.

Doing this check in ImportPlannedDisbursements means it's only done once
per import, whereas putting it in PlannedDisbursementHistory would mean
it was done for every forecast value.