Reject forecast imports containing unknown RODA identifiers
If we don't recognise any of the RODA identifiers in a forecast import,
we should fail the import.

I'm choosing to fail the entire thing if there are any errors. Even
though PlannedDisbursementHistory will not create duplicate records on
repeated input of the same data, I still think the interaction model is
simpler if the input succeeds or fails atomically.