Reject imports containing activities not related to the given report
When uploading forecast data to back-fill the last report, the developer
will specify which report all the new forecast data should go into.

It must be the case that for all planned disbursements, the report and
the activity they belong to must have the same fund and organisation.
That is, all the activities covered by a report should belong to the
same fund and organisation as that report.

Therefore we treat the presence of any activities not matching the given
report as an error and reject the whole import.