Script for importing forecasts into the last report
This script wraps the functionality of ImportPlannedDisbursements in a
command-line script. The arguments specify the input file, and which
report to assign the data to. After looking up the report, we ask the
user to confirm that's what they were expecting, as a safety check.

The arguments are:

    -f, --fund            The RODA identifier of the fund the imported
                          data relates to

    -o, --organisation    The name of the organisation that owns the
                          activities in the imported data

    -y, --year            The financial year of the report

    -q, --quarter         The financial quarter of the report

    -i, --input           The path to the file to import data from

If any errors happen during import, they are printed and the script
exits with 1. On success it prints nothing.

The use of a transaction inside ImportPlannedDisbursements means if any
exceptions occur during the import, any forecasts written up to that
point will be rolled back.