Import Channel of delivery code
For now, this field can only be imported from CSV.

We have decided not to add model validation for this field, until we have a better idea
about the way it is used and whether it is going to be integrated in the form steps or not.

Given that we can't rely on model validation to pick up on the field being blank,
we need to check for empty fields in the imported CSV, and reject such rows in the same way
that we reject any values that are not in code list.

We have removed a few lines of test because the specific error messages
are already being tested elsewhere. The activity may have multiple invalid attributes,
all we need to check for that specific test is that no other activities get updated either.

Co-authored-by: CristinaRO <cristina@dxw.com>