Don’t ignore blank columns when importing a new activity
Previously, the converter was ignoring any and all blank columns from the CSV.

Ignoring blank fields is a requirement for updating an existing activity.

However, when importing a new activity, we want to be able to validate the presence of certain attributes. These validations are performed by the `convert_attribute` methods.

I’ve added an optional argument to the converter, to allow updating an activity to continue as before.

Creating a new activity now processes all fields and passes them to the attribute converter, to do any validation specific to that attribute.