Handle zero values in PlannedDisbursementHistory
When a value of 0 is entered as the forecast for an activity and
quarter, whether by an end user or via a CSV import, that should be
interpreted as saying there is no planned spending for that time period.
PlannedDisbursementHistory needs to represent this fact in a way that
makes histoical sense.

At level A/B we do not pin history to reports. We create an `original`
entry when a forecast is first given for a quarter, we create a
`revised` entry the first time it's edited, and we modify the `revised`
entry after that. In this case, a zero value can be handled by:

- Ignoring it if no forecast for that quarter currently exists

- Storing it if we're creating/updating a `revised` entry, so that it
  supersedes the original value

That is, we never create `original` entries with value 0, but we _can_
create `revised` entries with value 0.

At level C/D, we pin the edit history to the reporting cycle. The latest
entry is updated if it belongs to the current report, otherwise we
create a new `revised` entry belonging to the current report, with the
new value, so that it supersedes all previous values.

At these levels the above rules apply, but there's one additional case.
If the latest entry belongs to the current report, and it's the only
existing entry for this forecast, we can delete it. That effectively
represents that there's no forecast for that time period, and there are
no prior values that need to be overridden.