Handle the existence of zero-valued forecasts in PlannedDisbursementOverview
The previous commit allows for the storage of 0 as the value of a
PlannedDisbursement, in order to represent that there used to be
forecasted spend for that activity/quarter, but this has been revised so
that no spending is now planned. We do this by storing 0 as the latest
revision value, not by deleting old history entries.

As 0 implies there is no forecasted spend, the following methods should
not include quarters for which the value is zero in their results:

- `PlannedDisbursementOverview#latest_values`
- `PlannedDisbursementOverview::Snapshot#all_quarters`

This has the effect that quarters whose latest value is 0 are no longer
displayed on the "financials" page for an activity.

The methods `value_for_report_quarter` and `value_for(financial_year:,
financial_quarter:)` in PlannedDisbursementOverview::Snapshot will still
return 0 when expected, because they are implemented by summing over the
rows returned by `all_quarters` that match the requested quarter; if
there are no matching rows then SUM() returns 0.