Import policy markers
This adds the policy marker fields for import. We want
to treat blank fields as “not_assessed”, but the way
the convertor works at the moment, we skip any
blank fields. I’ve got round this by adding a constant
called `ALLOWED_BLANK_FIELDS`. Rather than
just checking a field is present before converting it,
we additionally check is a field is in the
`ALLOWED_BLANK_FIELDS` constant too. This way
we can convert any fields that are left blank to read
as `not_assessed`