Adjust specs inline with code review comment
During code review it was noted that the two of the specs that changed
still read as asserting that an error was present, even though the
actual error message is covered elsewhere, we should include a check
that an error is present.

To ensure the error we are counting is the only error, I've adjusted the
setup to have a complete and valid set of attributes and then changed
the single attribute under test - this way we can be confident that
there will only be one single error.