Change Programme Status column to an integer
To allow us to make Programme Status into an enum, we need to make it into an
integer. As part of this migration, we need to create a new temporily named
column, go through any existing data and add the programme status as an integer
to the new column, then when all the data is backed up, remove the old column,
and remame the new column.