Update transactions to reflect actual disbursements
The definition of actuals/transactions has changed over time and we need
to update production data to reflect this. There are two parts to this
problem:

- Some transactions are included in the data which do not fit our
  definition, and they need to be removed.
- Some transactions are labelled incorrectly, and they need to be
  relabelled to fit our definition.

Any transactions of type 'incoming commitments' (code `11`) and
'outgoing commitments' (code `2`) should be deleted

All transactions in the database should be of type 'disbursement' code
`3`. To achieve this we need to change the type for 'expenditure' (code
`4`) and 'credit guarantee' (code `10`) to code `3`.