Link level C/D forecasts to reports
In the most recent deployment, part of the logic of the data migration
`20201124145240_link_forecasts_to_reports` failed to run, specifically
this section:

    level_c_or_d_forecasts.where(report_id: nil).each do |planned_disbursement|
      reports = historic_reports_for_activity(planned_disbursement.parent_activity)

      if reports.count == 1
        planned_disbursement.update!(report: reports.first)
      end
    end

The problem with this is the call to `update!` did not work, because
this migration ended up in a deployment where the PlannedDisbursement
model gained a new restriction on updates:

    attr_readonly :report_id

This makes the call to `update!` fail silently: it is not an error to
update a read-only column, the request is simply ignored. The main
instance-level alternative is `update_column`, which raises an exception
on read-only columns.

If we do want to actually update this column, we need to use the
relation-level `update_all` method. So this migration repeats the logic
of the partially-failed migration in a way that bypasses the read-only
restriction on `report_id`.