Transaction form and details view simplified
The current flow for reporting a transaction asks users to complete lots of
fields, not all of which are needed. We can save users time and effort by
streamlining the form.

Our proposed approach is to remove a number of fields that can be autopopulated
by RODA. This commit removes the fields that are not necessary in both the
`Add a transaction` form and the transaction details page.

I have also modified the specs to reflect these changes.