Make Staff::ActivitiesController show the latest planned disbursements
The `show` action in this controller was using a construct like
`PlannedDisbursement.where(parent_activity: activity)` to load the
planned disbursements. This would load all the historical versions of
forecasts for the activity, rather than the latest values which is
probably what we want.

We amend this to use `Activity#latest_planned_disbursements` so that
only the latest values are shown.